Add attrs.inspect()
