+++ src/attrs/__init__.py
@@ -4,6 +4,7 @@ from attr import (
     NOTHING,
     Attribute,
     AttrsInstance,
+    ClassProps,
     Converter,
     Factory,
     NothingType,
@@ -17,6 +18,7 @@ from attr import (
     fields_dict,
     frozen,
     has,
+    inspect,
     make_class,
     mutable,
     resolve_types,
@@ -31,6 +33,7 @@ __all__ = [
     "NOTHING",
     "Attribute",
     "AttrsInstance",
+    "ClassProps",
     "Converter",
     "Factory",
     "NothingType",
@@ -58,6 +61,7 @@ __all__ = [
     "filters",
     "frozen",
     "has",
+    "inspect",
     "make_class",
     "mutable",
     "resolve_types",
